new: Add support of dependency tracking mechanism
Generate *.d (dependency) files, which allows to track changes of the
file and recompile files that depends on the changed. Basically, it was
implemented to track changes of header files, so now when the header
changed - the build system will recompile all files that include it.

Small changes in rules for object files generation.

Changes in environment.mk mostly cosmetical, but also define test
directories and test executables.

Define .PHONY targets.

Signed-off-by: Vladyslav Tsymbal <vladyslav.tsymbal@globallogic.com>