new+ref: Add Texturing. Refactor old classes
Add support of texturing. For this feature the following changes has
been done:
- Add GLTexture class
- Add support of texturing inside Model class
- Edit a bit vertex and fragment shaders for example app`
- Modify Render mechanism

Fix builder pattern for Shader and Model classes. Previously, I used
to modify the state of the object in builder class, what is totally
wrong. Builder pattern should only be used in way of convenient passing
data to the class constructor.

Reorganise the structure and relations between Camera, Window,
RenderSystem and GlobalEngine classes. Now the GlobalEngine class
contains Camera, Window and RenderSystem instances. Moving this classes
to the GlobalEngine class, perhaps helps to follow single responsibility
rule for this classes.

Drop maths directory, because of it's redundancy. Now, the RenderSystem class
is responsible for this functionality. RenderSystem is able to create view and
projection matrixes with required params.

Remove RenderProperties class and use the according structure instead.

Use PIMPL idiom in case of GlobalEngine class. This helps to hide
implementation from users of the class and decrease compilation time.

Add local .vimrc for current project. This allows to compile project
with pressing F4 and open headers via "go forward" shortcut.

Other minor changes...

Signed-off-by: Vladyslav Tsymbal <vladyslav.tsymbal@globallogic.com>