fix: Fix the error with include path for target in Makefiles
The include paths, which are used in rules for creating GlobalEngine
target was not proper, because they have been overwritten with paths
for tests. So, that was the reason to update the build system itself.

Move common part of building executable binary file to binary.mk.
executable.mk and test.mk simply set up the environment and include
the binary.mk, which contains basic rules.

Use LOCAL_MODULE_ prefix for describing all neccessary variables for
creating the module. "local_" prefix is used in internal Makefiles.
The DEFAULT_ prefix is used for all major variables, which is commonly
used. PRIVATE_ prefix is used for variables, which is used only in
recipes.

Signed-off-by: Vladyslav Tsymbal <vladyslav.tsymbal@gmail.com>